Beta coefficients

Computes the beta (regression) coefficients, also called standardized coefficients or beta weights, to quantify the importance (the effect) of the predictors on the dependent variable in a multiple regression analysis where the variables are measured in different units.

beta.coef(object, ...)

# S3 method for default
beta.coef(object, complete = FALSE, ...)

Arguments

object

an object for which the extraction of model coefficients is meaningful.

...

further arguments passed to or from other methods.

complete

for the default (used for lm, etc) and aov methods: logical indicating if the full coefficient vector should be returned also in case of an over-determined system where some coefficients will be set to NA.

Details

The beta weights are the coefficient estimates resulting from a regression analysis where the underlying data have been standardized so that the variances of dependent and explanatory variables are equal to 1. Therefore, standardized coefficients are unitless and refer to how many standard deviations a dependent variable will change, per standard deviation increase in the predictor variable. See https://en.wikipedia.org/wiki/Standardized_coefficient or QuantPsyc::lm.beta.

Based on QuantPsyc::lm.beta.

See also

Examples

fit <- lm(fidelida ~ velocida + calidadp, hbat)
coef(fit)
#> (Intercept)    velocida    calidadp 
#>    1.334685    7.973594    3.336468 
beta.coef(fit)
#>  velocida  calidadp 
#> 0.6664119 0.5146444 
fit2 <- lm(scale(fidelida) ~ scale(velocida) + scale(calidadp), hbat)
coef(fit2)
#>     (Intercept) scale(velocida) scale(calidadp) 
#>   -4.566135e-16    6.664119e-01    5.146444e-01