HBAT data

A dataset containing observations of customers of the industrial distribution company HBAT. The variables can be classified into three groups: the first 6 (categorical) are shopper characteristics (data warehouse classification), variables 7 to 19 (numerical) measure shopper perceptions of HBAT and the last 5 are possible target variables (responses), the purchase outcomes.

Format

A data frame with 200 rows and 24 columns:

empresa

Customer ID.

tcliente

Customer Type. Length of time a particular customer has been buying from HBAT: Menos de 1 año = less than 1 year. De 1 a 5 años = between 1 and 5 years. Más de 5 años = longer than 5 years.

tindustr

Type of industry that purchases HBAT’s paper products: Revista = magazine industry, Periodico = newsprint industry.

tamaño

Employee size: Pequeña (<500) = small firm, fewer than 500 employees, Grande (>=500) = large firm, 500 or more employees.

region

Customer location: America del norte = USA/North America, Otros = outside North America.

distrib

Distribution System. How paper products are sold to customers: Indirecta = sold indirectly through a broker, Directa = sold directly.

calidadp

Product Quality. Perceived level of quality of HBAT’s paper products.

web

E-Commerce Activities/Web Site. Overall image of HBAT’s Web site, especially user-friendliness.

soporte

Technical Support. Extent to which technical support is offered to help solve product/service issues.

quejas

Complaint Resolution. Extent to which any complaints are resolved in a timely and complete manner.

publi

Advertising. Perceptions of HBAT’s advertising campaigns in all types of media.

producto

Product Line. Depth and breadth of HBAT’s product line to meet customer needs.

imgfvent

Salesforce Image. Overall image of HBAT’s salesforce.

precio

Competitive Pricing. Extent to which HBAT offers competitive prices.

garantia

Warranty and Claims. Extent to which HBAT stands behind its product/service warranties and claims.

nprod

New Products. Extent to which HBAT develops and sells new products.

facturac

Ordering and Billing. Perception that ordering and billing is handled efficiently and correctly.

flexprec

Price Flexibility. Perceived willingness of HBAT sales reps to negotiate price on purchases of paper products.

velocida

Delivery Speed. Amount of time it takes to deliver the paper products once an order has been confirmed.

satisfac

Customer satisfaction with past purchases from HBAT, measured on a 10-point graphic rating scale.

precomen

Likelihood of recommending HBAT to other firms as a supplier of paper products, measured on a 10-point graphic rating scale.

pcompra

Likelihood of purchasing paper products from HBAT in the future, measured on a 10-point graphic rating scale.

fidelida

Percentage of Purchases from HBAT. Percentage of the responding firm’s paper needs purchased from HBAT, measured on a 100-point percentage scale.

alianza

Perception of Future Relationship with HBAT. Extent to which the customer/respondent perceives his or her firm would engage in strategic alliance/partnership with HBAT: No = Would not consider. Si = Yes, would consider strategic alliance or partnership.

Source

Hair et al. (1998).

Details

For more details, consult the reference Hair et al. (1998).

References

Hair, J. F., Anderson, R. E., Tatham, R. L., y Black, W. (1998). Multivariate Data Analysis. Prentice Hall.

Examples

str(hbat)
#> 'data.frame':	200 obs. of  24 variables:
#>  $ empresa : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ tcliente: Factor w/ 3 levels "Menos de 1 año",..: 2 3 3 1 2 1 1 2 2 1 ...
#>  $ tindustr: Factor w/ 2 levels "Revista","Periodico": 1 2 1 2 1 2 2 1 2 1 ...
#>  $ tamaño  : Factor w/ 2 levels "Pequeña  (<500)",..: 2 1 2 2 2 1 2 2 2 2 ...
#>  $ region  : Factor w/ 2 levels "America del norte",..: 2 1 2 2 1 2 2 2 2 2 ...
#>  $ distrib : Factor w/ 2 levels "Indirecta","Directa": 2 1 2 1 2 1 1 1 1 1 ...
#>  $ calidadp: num  8.5 8.2 9.2 6.4 9 6.5 6.9 6.2 5.8 6.4 ...
#>  $ web     : num  3.9 2.7 3.4 3.3 3.4 2.8 3.7 3.3 3.6 4.5 ...
#>  $ soporte : num  2.5 5.1 5.6 7 5.2 3.1 5 3.9 5.1 5.1 ...
#>  $ quejas  : num  5.9 7.2 5.6 3.7 4.6 4.1 2.6 4.8 6.7 6.1 ...
#>  $ publi   : num  4.8 3.4 5.4 4.7 2.2 4 2.1 4.6 3.7 4.7 ...
#>  $ producto: num  4.9 7.9 7.4 4.7 6 4.3 2.3 3.6 5.9 5.7 ...
#>  $ imgfvent: num  6 3.1 5.8 4.5 4.5 3.7 5.4 5.1 5.8 5.7 ...
#>  $ precio  : num  6.8 5.3 4.5 8.8 6.8 8.5 8.9 6.9 9.3 8.4 ...
#>  $ garantia: num  4.7 5.5 6.2 7 6.1 5.1 4.8 5.4 5.9 5.4 ...
#>  $ nprod   : num  4.3 4 4.6 3.6 4.5 9.5 2.5 4.8 4.4 5.3 ...
#>  $ facturac: num  5 3.9 5.4 4.3 4.5 3.6 2.1 4.3 4.4 4.1 ...
#>  $ flexprec: num  5.1 4.3 4 4.1 3.5 4.7 4.2 6.3 6.1 5.8 ...
#>  $ velocida: num  3.7 4.9 4.5 3 3.5 3.3 2 3.7 4.6 4.4 ...
#>  $ satisfac: num  8.2 5.7 8.9 4.8 7.1 4.7 5.7 6.3 7 5.5 ...
#>  $ precomen: num  8 6.5 8.4 6 6.6 6.3 7.8 5.8 7.5 5.9 ...
#>  $ pcompra : num  8.4 7.5 9 7.2 9 6.1 7.2 7.7 8.2 6.7 ...
#>  $ fidelida: num  65 67 72 40 57 50 41 56 56 59 ...
#>  $ alianza : Factor w/ 2 levels "No","Si": 2 1 2 1 1 1 1 1 2 1 ...
#>  - attr(*, "variable.labels")= Named chr [1:24] "Empresa" "Tipo de cliente" "Tipo Industria" "Tamaño de la empresa" ...
#>   ..- attr(*, "names")= chr [1:24] "empresa" "tcliente" "tindustr" "tamaño" ...
#>  - attr(*, "codepage")= int 1252
as.data.frame(attr(hbat, "variable.labels"))
#>             attr(hbat, "variable.labels")
#> empresa                           Empresa
#> tcliente                  Tipo de cliente
#> tindustr                   Tipo Industria
#> tamaño               Tamaño de la empresa
#> region                             Región
#> distrib           Sistema de distribución
#> calidadp              Calidad de producto
#> web      Actividades comercio electrónico
#> soporte                   Soporte técnico
#> quejas               Resolución de quejas
#> publi                          Publicidad
#> producto               Línea de productos
#> imgfvent       Imagen de fuerza de ventas
#> precio                   Nivel de precios
#> garantia         Garantía y reclamaciones
#> nprod                    Nuevos productos
#> facturac            Encargo y facturación
#> flexprec          Flexibilidad de precios
#> velocida             Velocidad de entrega
#> satisfac            Nivel de satisfacción
#> precomen          Propensión a recomendar
#> pcompra              Propensión a comprar
#> fidelida      Porcentaje de compra a HBAT
#> alianza  Consideraría alianza estratégica
summary(hbat)
#>     empresa                 tcliente       tindustr               tamaño   
#>  Min.   :  1.00   Menos de 1 año:68   Revista  :100   Pequeña  (<500): 98  
#>  1st Qu.: 50.75   De 1 a 5 años :64   Periodico:100   Grande  (>=500):102  
#>  Median :100.50   Más de 5 años :68                                        
#>  Mean   :100.50                                                            
#>  3rd Qu.:150.25                                                            
#>  Max.   :200.00                                                            
#>                region         distrib       calidadp           web       
#>  America del norte: 81   Indirecta:108   Min.   : 5.000   Min.   :2.200  
#>  Otros            :119   Directa  : 92   1st Qu.: 6.700   1st Qu.:3.300  
#>                                          Median : 8.050   Median :3.700  
#>                                          Mean   : 7.894   Mean   :3.765  
#>                                          3rd Qu.: 9.100   3rd Qu.:4.100  
#>                                          Max.   :10.000   Max.   :5.700  
#>     soporte          quejas          publi          producto    
#>  Min.   :0.400   Min.   :1.500   Min.   :1.500   Min.   :2.300  
#>  1st Qu.:4.100   1st Qu.:4.500   1st Qu.:3.200   1st Qu.:4.800  
#>  Median :5.300   Median :5.400   Median :4.000   Median :5.700  
#>  Mean   :5.243   Mean   :5.367   Mean   :4.061   Mean   :5.815  
#>  3rd Qu.:6.425   3rd Qu.:6.200   3rd Qu.:4.925   3rd Qu.:6.900  
#>  Max.   :8.900   Max.   :8.700   Max.   :6.900   Max.   :8.400  
#>     imgfvent         precio          garantia         nprod      
#>  Min.   :2.500   Min.   : 3.700   Min.   :3.700   Min.   :1.200  
#>  1st Qu.:4.575   1st Qu.: 5.800   1st Qu.:5.400   1st Qu.:4.200  
#>  Median :5.000   Median : 7.150   Median :6.100   Median :5.200  
#>  Mean   :5.248   Mean   : 6.971   Mean   :6.048   Mean   :5.210  
#>  3rd Qu.:6.000   3rd Qu.: 8.325   3rd Qu.:6.600   3rd Qu.:6.225  
#>  Max.   :8.200   Max.   :10.000   Max.   :8.400   Max.   :9.500  
#>     facturac        flexprec        velocida        satisfac    
#>  Min.   :1.200   Min.   :2.300   Min.   :1.600   Min.   :4.700  
#>  1st Qu.:3.700   1st Qu.:3.575   1st Qu.:3.300   1st Qu.:6.000  
#>  Median :4.300   Median :4.250   Median :3.900   Median :7.050  
#>  Mean   :4.242   Mean   :4.464   Mean   :3.816   Mean   :6.952  
#>  3rd Qu.:4.700   3rd Qu.:5.325   3rd Qu.:4.325   3rd Qu.:7.900  
#>  Max.   :6.700   Max.   :7.500   Max.   :5.500   Max.   :9.900  
#>     precomen        pcompra         fidelida     alianza 
#>  Min.   :4.000   Min.   :4.300   Min.   :37.00   No:114  
#>  1st Qu.:6.100   1st Qu.:7.100   1st Qu.:51.00   Si: 86  
#>  Median :7.000   Median :7.600   Median :58.50           
#>  Mean   :6.952   Mean   :7.665   Mean   :58.10           
#>  3rd Qu.:7.650   3rd Qu.:8.200   3rd Qu.:65.25           
#>  Max.   :9.900   Max.   :9.900   Max.   :77.00