Wine taste data

A subset related to the white variant of the Portuguese "Vinho Verde" wine, containing physicochemical information (fixed.acidity, volatile.acidity, citric.acid, residual.sugar, chlorides, free.sulfur.dioxide, total.sulfur.dioxide , density, pH, sulphates and alcohol) and sensory (taste), which indicates the quality of the wine (it is considered good if the median of the wine quality evaluations, made by experts, who evaluated them between 0 = very bad and 10 = very excellent, is not less than 6.

Format

A data frame with 1,250 rows and 12 columns:

fixed.acidity

fixed acidity

volatile.acidity

volatile acidity

citric.acid

citric acid

residual.sugar

residual sugar

chlorides

chlorides

free.sulfur.dioxide

free sulfur dioxide

total.sulfur.dioxide

total sulfur dioxide

density

density

pH

pH

sulphates

sulphates

alcohol

alcohol

taste

factor with levels "good" and "bad" indicating the quality of the wine

Source

UCI Machine Learning Repository: http://archive.ics.uci.edu/dataset/186/wine+quality.

Details

For more details, consult http://www.vinhoverde.pt/en/ or the reference Cortez et al. (2009).

References

Cortez, P., Cerdeira, A., Almeida, F., Matos, T., & Reis, J. (2009). Modeling wine preferences by data mining from physicochemical properties. Decision Support Systems, 47(4), 547-553.

See also

Examples

winetaste <- winequality[, names(winequality)!="quality"]
winetaste$taste <- factor(winequality$quality < 6,
                      labels = c('good', 'bad')) # levels = c('FALSE', 'TRUE')
str(winetaste)
#> 'data.frame':	1250 obs. of  12 variables:
#>  $ fixed.acidity       : num  6.8 7.1 6.9 7.5 8.6 7.7 5.4 6.8 6.1 5.5 ...
#>  $ volatile.acidity    : num  0.37 0.24 0.32 0.23 0.36 0.28 0.59 0.16 0.28 0.28 ...
#>  $ citric.acid         : num  0.47 0.34 0.13 0.49 0.26 0.63 0.07 0.36 0.27 0.21 ...
#>  $ residual.sugar      : num  11.2 1.2 7.8 7.7 11.1 11.1 7 1.3 4.7 1.6 ...
#>  $ chlorides           : num  0.071 0.045 0.042 0.049 0.03 0.039 0.045 0.034 0.03 0.032 ...
#>  $ free.sulfur.dioxide : num  44 6 11 61 43.5 58 36 32 56 23 ...
#>  $ total.sulfur.dioxide: num  136 132 117 209 171 179 147 98 140 85 ...
#>  $ density             : num  0.997 0.991 0.996 0.994 0.995 ...
#>  $ pH                  : num  2.98 3.16 3.23 3.14 3.03 3.08 3.34 3.02 3.16 3.42 ...
#>  $ sulphates           : num  0.88 0.46 0.37 0.3 0.49 0.44 0.57 0.58 0.42 0.42 ...
#>  $ alcohol             : num  9.2 11.2 9.2 11.1 12 8.8 9.7 11.3 12.5 12.5 ...
#>  $ taste               : Factor w/ 2 levels "good","bad": 2 2 2 1 2 2 1 1 1 2 ...